New parser and lexer working.  Time to iron out lots of kinks.

+++ includes/ft_sh.h
@@ -58,9 +58,9 @@ typedef struct  		s_shell
 typedef struct			s_tokelist
 {
 	char				type[3];
-	size_t				len;
-	size_t				redir_prefix_fd;
-	size_t				redir_suffix_fd;
+	int					len;
+	int					redir_prefix_fd;
+	int					redir_suffix_fd;
 	char				*redir_suffix_file;
 	bool				redir_turn_off;
 	struct s_tokelist	*next;
